+++ db/migrate/20260705130623_create_fin_ops_transfers.rb
@@ -4,7 +4,7 @@ class CreateFinOpsTransfers < ActiveRecord::Migration[8.1]
       t.references :sender, null: false
       t.references :receiver, null: false
       t.uuid :public_id, null: false, default: "uuid_generate_v4()"
-      t.integer :amount_cents, null: false
+      t.bigint :amount_cents, null: false
 
       t.timestamps
     end
